update field codes

+++ hook-debug/accept/accept.c
@@ -3,11 +3,11 @@
  */
 #include "../hookapi.h"
 
-int64_t cbak(int64_t reserved) { 
+int64_t cbak(uint32_t reserved) { 
     return 0;
 }
 
-int64_t hook(int64_t reserved ) {
+int64_t hook(uint32_t reserved ) {
 
     TRACESTR("Accept.c: Called.");
     accept (0,0,0);
